requirements.txt
fastapi
uvicorn
python-dotenv
openai
PyMuPDF
numpy
easyocr
pdf2image
jinja2
os
fitz